requirements.txt
selenium
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
coverage==5.0.3
Flask==1.1.1
Flask-Limiter==1.1.0
google-api-core==1.15.0
google-auth==1.10.0
googleapis-common-protos==1.6.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
limits==1.4.1
MarkupSafe==1.1.1
opencensus==0.7.6
opencensus-context==0.1.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyJWT==1.7.1
pypyodbc
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wikipedia
python-aiml
aiml

